[OWL-792][agent] share common definition of MetricValueExtend
