+++ modules/agent/g/mq.go
@@ -3,6 +3,7 @@ package g
 import (
 	"encoding/json"
 	"fmt"
+	"github.com/Cepave/open-falcon-backend/common/model"
 	log "github.com/Sirupsen/logrus"
 	"github.com/streadway/amqp"
 )
@@ -13,7 +14,7 @@ func logOnError(err error, msg string) {
 	}
 }
 
-func SendToMQ(metrics []*MetricValueExtend) {
+func SendToMQ(metrics []*model.MetricValueExtend) {
 	if len(metrics) == 0 {
 		return
 	}
